requirements.txt
fastapi
uvicorn
python-telegram-bot
jinja2
python-multipart
httpx
qrcode
pillow
python-dotenv
